Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-time-serialize-expected.txt


FAIL e.style['transition-delay'] = "min(1ms)" should set the property value assert_equals: serialization should be canonical expected "min(1ms)" but got "calc(min(1ms))"
FAIL e.style['transition-delay'] = "min(1s)" should set the property value assert_equals: serialization should be canonical expected "min(1s)" but got "calc(min(1s))"
FAIL e.style['transition-delay'] = "max(1ms)" should set the property value assert_equals: serialization should be canonical expected "max(1ms)" but got "calc(max(1ms))"
FAIL e.style['transition-delay'] = "max(1s)" should set the property value assert_equals: serialization should be canonical expected "max(1s)" but got "calc(max(1s))"
FAIL e.style['transition-delay'] = "min(1ms, 2ms, 3ms)" should set the property value assert_equals: serialization should be canonical expected "min(1ms, 2ms, 3ms)" but got "calc(min(1ms, 2ms, 3ms))"
FAIL e.style['transition-delay'] = "min(3ms, 2ms, 1ms)" should set the property value assert_equals: serialization should be canonical expected "min(3ms, 2ms, 1ms)" but got "calc(min(3ms, 2ms, 1ms))"
FAIL e.style['transition-delay'] = "max(1ms, 2ms, 3ms)" should set the property value assert_equals: serialization should be canonical expected "max(1ms, 2ms, 3ms)" but got "calc(max(1ms, 2ms, 3ms))"
FAIL e.style['transition-delay'] = "max(3ms, 2ms, 1ms)" should set the property value assert_equals: serialization should be canonical expected "max(3ms, 2ms, 1ms)" but got "calc(max(3ms, 2ms, 1ms))"
FAIL e.style['transition-delay'] = "min(1000ms, 1s)" should set the property value assert_equals: serialization should be canonical expected "min(1000ms, 1s)" but got "calc(min(1000ms, 1s))"
FAIL e.style['transition-delay'] = "min(1s, 1000ms)" should set the property value assert_equals: serialization should be canonical expected "min(1s, 1000ms)" but got "calc(min(1s, 1000ms))"
FAIL e.style['transition-delay'] = "max(1000ms, 1s)" should set the property value assert_equals: serialization should be canonical expected "max(1000ms, 1s)" but got "calc(max(1000ms, 1s))"
FAIL e.style['transition-delay'] = "max(1s, 1000ms)" should set the property value assert_equals: serialization should be canonical expected "max(1s, 1000ms)" but got "calc(max(1s, 1000ms))"
FAIL e.style['transition-delay'] = "calc(min(1s) + min(2s))" should set the property value assert_equals: serialization should be canonical expected "calc(min(1s) + min(2s))" but got "calc(3ms)"
FAIL e.style['transition-delay'] = "calc(min(2s) + min(1s))" should set the property value assert_equals: serialization should be canonical expected "calc(min(2s) + min(1s))" but got "calc(3ms)"
FAIL e.style['transition-delay'] = "calc(max(1s) + max(2s))" should set the property value assert_equals: serialization should be canonical expected "calc(max(1s) + max(2s))" but got "calc(3ms)"
FAIL e.style['transition-delay'] = "calc(max(2s) + max(1s))" should set the property value assert_equals: serialization should be canonical expected "calc(max(2s) + max(1s))" but got "calc(3ms)"
FAIL e.style['transition-delay'] = "calc(1s + min(2s))" should set the property value assert_equals: serialization should be canonical expected "calc(1s + min(2s))" but got "calc(1002ms)"
FAIL e.style['transition-delay'] = "calc(min(2s) + 1s)" should set the property value assert_equals: serialization should be canonical expected "calc(1s + min(2s))" but got "calc(1002ms)"
FAIL e.style['transition-delay'] = "calc(1s + max(2s))" should set the property value assert_equals: serialization should be canonical expected "calc(1s + max(2s))" but got "calc(1002ms)"
FAIL e.style['transition-delay'] = "calc(max(2s) + 1s)" should set the property value assert_equals: serialization should be canonical expected "calc(1s + max(2s))" but got "calc(1002ms)"
